+++ src/numbers.ts
@@ -33,16 +33,18 @@ export const integer: Decoder<number> = number.refine(
 /**
  * Accepts only non-negative (zero or positive) finite numbers.
  */
-export const positiveNumber: Decoder<number> = number
-  .refine((n) => n >= 0, 'Number must be positive')
-  .transform(Math.abs); // Just here to handle the -0 case
+export const positiveNumber: Decoder<number> = number.refine(
+  (n) => n >= 0 && !Object.is(n, -0),
+  'Number must be positive',
+);
 
 /**
  * Accepts only non-negative (zero or positive) finite whole numbers.
  */
-export const positiveInteger: Decoder<number> = integer
-  .refine((n) => n >= 0, 'Number must be positive')
-  .transform(Math.abs); // Just here to handle the -0 case
+export const positiveInteger: Decoder<number> = integer.refine(
+  (n) => n >= 0 && !Object.is(n, -0),
+  'Number must be positive',
+);
 
 /**
  * Accepts any valid ``bigint`` value.